Polish, validation, fixes, css
