+++ src/onegov/org/locale/de_CH/LC_MESSAGES/onegov.org.po
@@ -2,7 +2,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE 1.0\n"
-"POT-Creation-Date: 2026-07-01 14:24+0200\n"
+"POT-Creation-Date: 2026-07-01 16:45+0200\n"
 "PO-Revision-Date: 2022-03-15 10:21+0100\n"
 "Last-Translator: Marc Sommerhalder <marc.sommerhalder@seantis.ch>\n"
 "Language-Team: German\n"
@@ -1942,6 +1942,12 @@ msgstr "Stornierung von Reservationen erlauben"
 
 msgid "Cancellation"
 msgstr "Stornierung"
+
+msgid ""
+"An e-mail reply address is required when cancellation requests are enabled."
+msgstr ""
+"Eine E-Mail-Antwortadresse ist erforderlich, wenn Stornierungsanfragen "
+"aktiviert sind."
 
 msgid ""
 "Failed to retrieve the doors from the dormakaba API please make sure your "
@@ -5171,6 +5177,12 @@ msgstr ""
 msgid "${author} wrote"
 msgstr "${author} schrieb"
 
+msgid "The following reservation remains active:"
+msgstr "Die folgende Reservation bleibt aktiv:"
+
+msgid "The following reservations remain active:"
+msgstr "Die folgenden Reservationen bleiben aktiv:"
+
 msgid "Check request status"
 msgstr "Anfragestatus überprüfen"
 
@@ -6238,9 +6250,6 @@ msgstr "Dann fügen Sie der Anfrage eine Nachricht hinzu."
 
 msgid "New messages have been disabled because the ticket has been closed."
 msgstr "Nachrichten wurden deaktiviert, da das Ticket geschlossen wurde."
-
-msgid "Would you like to cancel your reservation?"
-msgstr "Möchten Sie Ihre Reservation stornieren?"
 
 msgid ""
 "Enable notifications about new tickets. Only works when being logged in and "
@@ -8067,6 +8076,14 @@ msgstr "Ihre Anfrage wurde bereits geschlossen"
 msgid "Your message has been received"
 msgstr "Ihre Nachricht wurde übermittelt"
 
+msgid "Cancellation requests are not enabled for this resource."
+msgstr "Stornierungsanfragen sind für diese Ressource nicht aktiviert."
+
+msgid ""
+"All reservations for this ticket have already been cancelled."
+msgstr ""
+"Alle Reservationen für dieses Ticket wurden bereits storniert."
+
 msgid "A cancellation request has already been submitted."
 msgstr "Eine Stornierungsanfrage wurde bereits eingereicht."
 
@@ -8175,6 +8192,9 @@ msgstr "Der Benutzer wurde erfolgreich erstellt"
 
 msgid "Please enter your e-mail address in order to continue"
 msgstr "Bitte geben Sie ihre E-Mail Adresse ein um fortzufahren"
+
+#~ msgid "Would you like to cancel your reservation?"
+#~ msgstr "Möchten Sie Ihre Reservation stornieren?"
 
 #~ msgid "Cancellation requested for"
 #~ msgstr "Stornierungsanfrage für"
