+++ src/onegov/org/templates/mail_cancellation_accepted.pt
@@ -17,6 +17,14 @@
         <p><b>${resource.title}</b></p>
         <tal:b metal:use-macro="layout.macros['reservations']"></tal:b>
 
+        <tal:b tal:condition="remaining_reservations|nothing">
+            <p tal:condition="python:len(remaining_reservations) == 1" i18n:translate>The following reservation remains active:</p>
+            <p tal:condition="python:len(remaining_reservations) > 1" i18n:translate>The following reservations remain active:</p>
+            <p><b>${resource.title}</b></p>
+            <tal:b tal:define="reservations remaining_reservations"
+                   metal:use-macro="layout.macros['reservations']"></tal:b>
+        </tal:b>
+
         <p>
             <a href="${request.link(ticket, 'status')}" i18n:translate>Check request status</a>
         </p>
