+++ src/onegov/org/theme/styles/org.scss
@@ -4503,7 +4503,8 @@ ul.search-results {
 /*
     Ticket summary customisations
 */
-.ticket-summary {
+.ticket-summary,
+.ticket-cancellation-request {
     margin-bottom: 1.5rem;
 
     > div {
@@ -5574,6 +5575,7 @@ h2 + div.timeline {
     }
 
     .ticket-summary h2,
+    .ticket-cancellation-request h2,
     .ticket-timeline h2 {
         font-weight: bold;
     }
@@ -5582,7 +5584,8 @@ h2 + div.timeline {
         display: none;
     }
 
-    .ticket-summary {
+    .ticket-summary,
+    .ticket-cancellation-request {
         margin-bottom: 0;
 
         * {
