+++ src/onegov/town6/templates/mail_rejected_reservation_notification.pt
@@ -31,7 +31,7 @@
         </tal:b>
 
         <p i18n:translate>
-            See Ticket <a href="${request.link(model)}" i18n:name="link">${ticket_reference}</a> for additional information.
+            See Ticket <a href="${request.link(ticket, 'status')}" i18n:name="link">${ticket_reference}</a> for additional information.
         </p>
         <p i18n:translate>Have a great day!</p>
     </tal:b>
